% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects.R
\name{get_objects}
\alias{get_objects}
\title{Retrieve objects for a given predicate of a LINDAS resource}
\usage{
get_objects(uri, predicate, lang_pref = c("en", "de", "fr", "it", "rm"))
}
\arguments{
\item{uri}{Resource URI.}

\item{predicate}{Predicate URI.}

\item{lang_pref}{Preferred order of language codes used to resolve
  multilingual literals. Supported language codes typically include
  \code{"de"} (German), \code{"fr"} (French), \code{"it"} (Italian),
  \code{"rm"} (Romansh), and \code{"en"} (English).

  If a character vector is supplied, literals sharing the same predicate
  are reduced to a single value according to the order specified in
  \code{lang_pref}. The first available value in the preferred language
  order is selected. If none of the preferred languages are available,
  the first non-missing value is returned.

  If \code{NULL} is supplied, all language-tagged values returned by the
  SPARQL endpoint are kept.}
}
\value{
A tibble with subject, predicate, object, object language, and
  object datatype.
}
\description{
\code{get_objects} retrieves all objects for a given predicate of a subject
URI in LINDAS.
}
\details{
If the objects are multilingual literals, they can be reduced according to
the preferred language order given in \code{lang_pref}.
}
\examples{
\dontrun{
get_objects(
  "https://communication.ld.admin.ch/ofcom/srg_revenue_structure/10",
  "http://schema.org/name"
)
}
}
