% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicates.R
\name{get_predicates}
\alias{get_predicates}
\title{Retrieve predicates for a given LINDAS resource}
\usage{
get_predicates(uri)
}
\arguments{
\item{uri}{Resource URI.}
}
\value{
A tibble with one column, \code{p}, containing the predicates used
  for the given resource.
}
\description{
\code{get_predicates} retrieves all predicates used for a given subject URI
in LINDAS.
}
\examples{
\dontrun{
get_predicates("https://communication.ld.admin.ch/ofcom/srg_revenue_structure/10")
}
}
