% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_inbound.R
\name{get_resource_inbound}
\alias{get_resource_inbound}
\title{Retrieve inbound triples for a given LINDAS resource}
\usage{
get_resource_inbound(uri)
}
\arguments{
\item{uri}{Resource URI.}
}
\value{
A tibble with subject, predicate, and object for all inbound triples.
}
\description{
\code{get_resource_inbound} retrieves all triples in which the given URI
appears in object position.
}
\details{
If the subject has multilingual literals attached, no language reduction is
applied because this function returns inbound references rather than
properties of a single subject.
}
\examples{
\dontrun{
get_resource_inbound("https://ld.admin.ch/canton/21")
}
}
