% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lindasR.R
\docType{package}
\name{lindasR-package}
\alias{lindasR}
\alias{lindasR-package}
\title{lindasR: Interface to the Swiss Federal Linked Data Service}
\description{
Provides convenient access to the Swiss Federal Linked Data Service
(LINDAS) via its SPARQL endpoint.
}
\details{
See the README on
\href{https://github.com/zumbov2/lindasR#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ld.admin.ch}
  \item Report bugs at \url{https://github.com/zumbov2/lindas/issues}
}

}
\author{
\strong{Maintainer}: David Zumbach \email{david.zumbach@gfzb.ch}

}
\keyword{internal}
