% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_lindas.R
\name{query_lindas}
\alias{query_lindas}
\title{Query the LINDAS SPARQL endpoint}
\usage{
query_lindas(query)
}
\arguments{
\item{query}{A SPARQL query as a character string.}
}
\value{
A tibble containing the query result. If the query returns no rows,
  an empty tibble is returned.
}
\description{
\code{query_lindas} sends a SPARQL query to the LINDAS endpoint and returns
the result as a tibble.
}
